# app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: whisperX REST service
  description: '

    # whisperX REST Service


    Welcome to the whisperX RESTful API! This API provides a suite of audio processing
    services to enhance and analyze your audio content.


    ## Documentation:


    For detailed information on request and response formats, consult the [WhisperX
    Documentation](https://github.com/m-bain/whisperX).


    ## Services:


    Speech-2-Text provides a suite of audio processing services to enhance and analyze
    your audio content. The following services are available:


    1. Transcribe: Transcribe an audio/video  file into text.

    2. Align: Align the transcript to the audio/video file.

    3. Diarize: Diarize an audio/video file into speakers.

    4. Combine Transcript and Diarization: Combine the transcript and diarization
    results.


    ## Supported file extensions:

    AUDIO_EXTENSIONS = {''.amr'', ''.m4a'', ''.awb'', ''.aac'', ''.oga'', ''.wav'',
    ''.mp3'', ''.ogg'', ''.wma''}


    VIDEO_EXTENSIONS = {''.wmv'', ''.mkv'', ''.mp4'', ''.avi'', ''.mov''}


    '
  version: 0.0.1
paths:
...
  /task/{identifier}:
    get:
      tags:
      - Tasks Management
      summary: Get Transcription Status
      operationId: get_transcription_status_task__identifier__get
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}/delete:
    delete:
      tags:
      - Tasks Management
      summary: Delete Task
      operationId: delete_task_task__identifier__delete_delete
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Metadata:
      properties:
        task_type:
          type: string
          title: Task Type
        task_params:
          anyOf:
          - type: object
          - type: 'null'
          title: Task Params
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration
        audio_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Audio Duration
      type: object
      required:
      - task_type
      - task_params
      - language
      - file_name
      - url
      - duration
      title: Metadata
    Response:
      properties:
        identifier:
          type: string
          title: Identifier
        message:
          type: string
          title: Message
      type: object
      required:
      - identifier
      - message
      title: Response
    Result:
      properties:
        status:
          type: string
          title: Status
        result:
          title: Result
        metadata:
          $ref: '#/components/schemas/Metadata'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - status
      - result
      - metadata
      - error
      title: Result
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...